Normエラー対応
