+++ srcs/utils/envp.c
@@ -6,7 +6,7 @@
 /*   By: hnagasak <hnagasak@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/27 01:45:42 by hnagasak          #+#    #+#             */
-/*   Updated: 2024/03/20 22:09:19 by hnagasak         ###   ########.fr       */
+/*   Updated: 2024/03/28 09:18:07 by hnagasak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,39 +69,4 @@ size_t	get_argc(char *argv[])
 int	is_env_key_char(char c)
 {
 	return (ft_isalnum(c) || c == '_');
-}
-
-#include <sys/stat.h>
-
-int	is_executable_file(char *path)
-{
-	struct stat	buf;
-
-	if (stat(path, &buf) == -1)
-		return (0);
-	if (S_ISREG(buf.st_mode) && (buf.st_mode & S_IXUSR))
-		return (1);
-	return (0);
-}
-
-int is_updatable_file(char *path)
-{
-	struct stat	buf;
-
-	if (stat(path, &buf) == -1)
-		return (0);
-	if (S_ISREG(buf.st_mode) && (buf.st_mode & S_IWUSR))
-		return (1);
-	return (0);
-}
-
-int	is_directory(const char *path)
-{
-	struct stat	statbuf;
-
-	if (stat(path, &statbuf) != 0)
-	{
-		return (0); // エラーがあれば、ディレクトリではないと仮定
-	}
-	return (S_ISDIR(statbuf.st_mode));
 }
